TST: Check ndarray dtype by default

+++ pandas/tests/series/test_analytics.py
@@ -1358,13 +1358,13 @@ class TestSeriesAnalytics(TestData, tm.TestCase):
         tm.assert_equal(r, e)
 
         r = s.searchsorted([30])
-        e = np.array([2])
+        e = np.array([2], dtype=np.int64)
         tm.assert_numpy_array_equal(r, e)
 
     def test_searchsorted_numeric_dtypes_vector(self):
         s = Series([1, 2, 90, 1000, 3e9])
         r = s.searchsorted([91, 2e6])
-        e = np.array([3, 4])
+        e = np.array([3, 4], dtype=np.int64)
         tm.assert_numpy_array_equal(r, e)
 
     def test_search_sorted_datetime64_scalar(self):
@@ -1378,14 +1378,14 @@ class TestSeriesAnalytics(TestData, tm.TestCase):
         s = Series(pd.date_range('20120101', periods=10, freq='2D'))
         v = [pd.Timestamp('20120102'), pd.Timestamp('20120104')]
         r = s.searchsorted(v)
-        e = np.array([1, 2])
+        e = np.array([1, 2], dtype=np.int64)
         tm.assert_numpy_array_equal(r, e)
 
     def test_searchsorted_sorter(self):
         # GH8490
         s = Series([3, 1, 2])
         r = s.searchsorted([0, 3], sorter=np.argsort(s))
-        e = np.array([0, 2])
+        e = np.array([0, 2], dtype=np.int64)
         tm.assert_numpy_array_equal(r, e)
 
     def test_is_unique(self):
